refactor: reorganization following #131
