+++ pyTMD/eop.py
@@ -2,7 +2,7 @@
 u"""
 eop.py
 Written by Tyler Sutterley (11/2022)
-Utilities for maintaining Earth Orientation Parameter (EOP) files
+Utilities for maintaining and calculating Earth Orientation Parameters (EOP)
 
 PYTHON DEPENDENCIES:
     numpy: Scientific Computing Tools For Python
@@ -243,3 +243,152 @@ def update_finals_file(username=None, password=None, verbose=False, mode=0o775):
         pass
     else:
         return
+
+# read table of mean pole values, calculate angular coordinates at epoch
+def iers_mean_pole(input_file, input_epoch, version, **kwargs):
+    """
+    Calculates the angular coordinates of the IERS Conventional Mean Pole (CMP)
+
+    Parameters
+    ----------
+    input_file: str
+        Full path to mean-pole.tab file provided by IERS
+    input_epoch: float
+        Dates for the angular coordinates of the Conventional Mean Pole
+        in decimal years
+    version: str
+        Year of the conventional model
+    fill_value: float, default np.nan
+        Value for invalid flags
+
+    Returns
+    -------
+    x: float
+        Angular coordinate x of conventional mean pole
+    y: float
+        Angular coordinate y of conventional mean pole
+    flag: bool
+        epoch is valid for version and version number is valid
+
+    References
+    ----------
+    .. [1] Petit, G. and Luzum, B. (eds.), IERS Conventions (2010),
+        IERS Technical Note No. 36, BKG (2010)
+    """
+    # set default keyword arguments
+    kwargs.setdefault('fill_value', np.nan)
+    # verify IERS model version
+    assert version in ('2003','2010','2015'), "Incorrect IERS model version"
+    # read mean pole file
+    table = np.loadtxt(os.path.expanduser(input_file))
+    # reduce to 1971 to end date
+    ii, = np.nonzero(table[:,0] >= 1971)
+    table = np.copy(table[ii,:])
+    # reduce to yearly values
+    jj, = np.nonzero((table[:,0] % 1) == 0.0)
+    table = np.copy(table[jj,:])
+    end_time = table[-1,0] + 0.2
+    # final shape of the table
+    nrows, ncols = np.shape(table)
+    # allocate for output arrays
+    x = np.full_like(input_epoch, kwargs['fill_value'])
+    y = np.full_like(input_epoch, kwargs['fill_value'])
+    flag = np.zeros_like(input_epoch, dtype=bool)
+    for t,epoch in enumerate(input_epoch):
+        # Conventional mean pole model in IERS Conventions 2003
+        if (version == '2003') and (epoch >= 1975) and (epoch < 2004):
+            x[t] = 0.054 + 0.00083*(epoch-2000.0)
+            y[t] = 0.357 + 0.00395*(epoch-2000.0)
+            flag[t] = True
+        # Conventional mean pole model in IERS Conventions 2010
+        elif (version == '2010') and (epoch >= 1975) and (epoch < 2011):
+            dx = epoch-2000.0
+            if (dx < 10.0):
+                x[t] = 0.055974 + 1.8243e-3*dx + 1.8413e-4*dx**2 + 7.024e-6*dx**3
+                y[t] = 0.346346 + 1.7896e-3*dx + 1.0729e-4*dx**2 + 0.908e-6*dx**3
+            else:
+                x[t] = 0.023513 + 0.0076141*dx
+                y[t] = 0.358891 - 0.0006287*dx
+            flag[t] = True
+        # Conventional mean pole model in IERS Conventions 2015
+        # must be below maximum valid date within file (e.g. 2015.2 for 2015)
+        elif (version == '2015') and (epoch >= 1975) and (epoch < end_time):
+            # find epoch within mean pole table
+            i = 1
+            j = nrows+1
+            while (j > (i+1)):
+                k = (i+j)//2
+                if (epoch < table[k,0]):
+                    j = k
+                else:
+                    i = k
+            # calculate differential from point in table
+            dx = epoch - table[i,0]
+            if (i == (nrows-1)):
+                x[t] = table[i,1] + dx*(table[nrows-1,1]-table[nrows-2,1])
+                y[t] = table[i,2] + dx*(table[nrows-1,1]-table[nrows-2,2])
+            else:
+                x[t] = table[i,1] + dx*(table[i+1,1]-table[i,1])
+                y[t] = table[i,2] + dx*(table[i+1,2]-table[i,2])
+            flag[t] = True
+    # return mean pole values
+    return (x, y, flag)
+
+# PURPOSE: read daily earth orientation parameters (EOP) file from IERS
+def iers_daily_EOP(input_file):
+    """
+    Read daily earth orientation parameters (EOP) file from IERS
+
+    Parameters
+    ----------
+    input_file: str
+        full path to IERS EOP "finals" file
+
+    Returns
+    -------
+    MJD: float
+        modified Julian date of EOP measurements
+    x: float
+        Angular coordinate x [arcsec]
+    y: float
+        Angular coordinate y [arcsec]
+
+    References
+    ----------
+    [Petit2010] G. Petit, and B. Luzum, (eds.), IERS Conventions (2010),
+        IERS Technical Note No. 36, BKG (2010)
+    """
+    # tilde-expansion of input file
+    input_file = os.path.expanduser(input_file)
+    # check that IERS finals file is accessible
+    if not os.access(input_file, os.F_OK):
+        raise FileNotFoundError(input_file)
+    # read data file splitting at line breaks
+    with open(input_file, mode='r', encoding='utf8') as f:
+        file_contents = f.read().splitlines()
+    # number of data lines
+    n_lines = len(file_contents)
+    dinput = {}
+    dinput['MJD'] = np.zeros((n_lines))
+    dinput['x'] = np.zeros((n_lines))
+    dinput['y'] = np.zeros((n_lines))
+    # for each line in the file
+    flag = 'I'
+    counter = 0
+    while (flag == 'I'):
+        line = file_contents[counter]
+        i = 2+2+2+1; j = i+8
+        dinput['MJD'][counter] = np.float64(line[i:j])
+        i = j+1
+        flag = line[i]
+        i += 2; j = i+9
+        dinput['x'][counter] = np.float64(line[i:j])
+        i = j+10; j = i+9
+        dinput['y'][counter] = np.float64(line[i:j])
+        counter += 1
+    # reduce to data values
+    dinput['MJD'] = dinput['MJD'][:counter]
+    dinput['x'] = dinput['x'][:counter]
+    dinput['y'] = dinput['y'][:counter]
+    # return the date, flag and polar motion values
+    return dinput
