+++ pyTMD/predict_tide.py
@@ -45,11 +45,10 @@ UPDATE HISTORY:
     Updated 07/2018: added option to use GSFC GOT nodal corrections
     Updated 09/2017: Rewritten in Python
 """
-import numpy as np
-from pyTMD.load_constituent import load_constituent
-from pyTMD.load_nodal_corrections import load_nodal_corrections
+import warnings
+import pyTMD.predict
 
-def predict_tide(t, hc, constituents, deltat=0.0, corrections='OTIS'):
+def predict_tide(*args, **kwargs):
     """
     Predict tides at a single time using harmonic constants
 
@@ -76,28 +75,8 @@ def predict_tide(t, hc, constituents, deltat=0.0, corrections='OTIS'):
     .. [1] Egbert and Erofeeva, "Efficient Inverse Modeling of Barotropic
         Ocean Tides", Journal of Atmospheric and Oceanic Technology, (2002).
     """
-
-    # number of points and number of constituents
-    npts,nc = np.shape(hc)
-    # load the nodal corrections
-    # convert time to Modified Julian Days (MJD)
-    pu,pf,G = load_nodal_corrections(t + 48622.0, constituents,
-        deltat=deltat, corrections=corrections)
-    # allocate for output tidal elevation
-    ht = np.ma.zeros((npts))
-    ht.mask = np.zeros((npts),dtype=bool)
-    # for each constituent
-    for k,c in enumerate(constituents):
-        if corrections in ('OTIS','ATLAS','ESR','netcdf'):
-            # load parameters for each constituent
-            amp,ph,omega,alpha,species = load_constituent(c)
-            # add component for constituent to output tidal elevation
-            th = omega*t*86400.0 + ph + pu[0,k]
-        elif corrections in ('GOT','FES'):
-            th = G[0,k]*np.pi/180.0 + pu[0,k]
-        # sum over all tides
-        ht.data[:] += pf[0,k]*hc.real[:,k]*np.cos(th) - \
-            pf[0,k]*hc.imag[:,k]*np.sin(th)
-        ht.mask[:] |= (hc.real.mask[:,k] | hc.imag.mask[:,k])
-    # return the tidal elevation after removing singleton dimensions
-    return np.squeeze(ht)
+    # raise warnings for deprecation of module
+    warnings.filterwarnings("always")
+    warnings.warn("Deprecated. Please use pyTMD.predict instead",DeprecationWarning)
+    # call renamed version to not break workflows
+    return pyTMD.predict.map(*args, **kwargs)
