+++ pyTMD/read_GOT_model.py
@@ -1,6 +1,8 @@
 #!/usr/bin/env python
 u"""
-read_GOT_model.py (11/2022)
+read_GOT_model.py
+Written by Tyler Sutterley (12/2022)
+
 Reads files for Richard Ray's Global Ocean Tide (GOT) models and makes initial
     calculations to run the tide program
 Includes functions to extract tidal harmonic constants out of a tidal model for
@@ -39,6 +41,7 @@ PROGRAM DEPENDENCIES:
     nearest_extrap.py: nearest-neighbor extrapolation of data to coordinates
 
 UPDATE HISTORY:
+    Updated 12/2022: refactored tide read programs under io
     Updated 11/2022: use f-strings for formatting verbose or ascii output
     Updated 05/2022: reformat arguments to extract_GOT_constants definition
         changed keyword arguments to camel case
@@ -124,6 +127,10 @@ def extract_GOT_constants(ilon, ilat, model_files=None, **kwargs):
     constituents: list
         list of model constituents
     """
+    # raise warnings for deprecation of module
+    warnings.filterwarnings("always")
+    warnings.warn("Deprecated. Please use pyTMD.io instead",DeprecationWarning)
+
     # set default keyword arguments
     kwargs.setdefault('method', 'spline')
     kwargs.setdefault('extrapolate', False)
