+++ pyTMD/read_iers_EOP.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 u"""
 read_iers_EOP.py
-Written by Tyler Sutterley (04/2022)
+Written by Tyler Sutterley (12/2022)
 Provides the daily earth orientation parameters (EOP) from IERS
     http://www.usno.navy.mil/USNO/earth-orientation/eo-products/weekly
 Data format: http://maia.usno.navy.mil/ser7/readme.finals
@@ -24,6 +24,7 @@ REFERENCE:
         IERS Technical Note No. 36, BKG (2010)
 
 UPDATE HISTORY:
+    Updated 12/2022: refactor daily orientation as part of eop module
     Updated 04/2022: updated docstrings to numpy documentation format
         include utf-8 encoding in reads to be windows compliant
     Updated 07/2021: added check that IERS finals file is accessible
@@ -31,9 +32,8 @@ UPDATE HISTORY:
     Updated 07/2020: added function docstrings
     Written 09/2017
 """
-import os
-import re
-import numpy as np
+import warnings
+import pyTMD.eop
 
 # calculates the daily earth orientation parameters (EOP) from IERS
 def read_iers_EOP(input_file):
@@ -59,37 +59,8 @@ def read_iers_EOP(input_file):
     [Petit2010] G. Petit, and B. Luzum, (eds.), IERS Conventions (2010),
         IERS Technical Note No. 36, BKG (2010)
     """
-    # tilde-expansion of input file
-    input_file = os.path.expanduser(input_file)
-    # check that IERS finals file is accessible
-    if not os.access(input_file, os.F_OK):
-        raise FileNotFoundError(input_file)
-    # read data file splitting at line breaks
-    with open(input_file, mode='r', encoding='utf8') as f:
-        file_contents = f.read().splitlines()
-    # number of data lines
-    n_lines = len(file_contents)
-    dinput = {}
-    dinput['MJD'] = np.zeros((n_lines))
-    dinput['x'] = np.zeros((n_lines))
-    dinput['y'] = np.zeros((n_lines))
-    # for each line in the file
-    flag = 'I'
-    counter = 0
-    while (flag == 'I'):
-        line = file_contents[counter]
-        i = 2+2+2+1; j = i+8
-        dinput['MJD'][counter] = np.float64(line[i:j])
-        i = j+1
-        flag = line[i]
-        i += 2; j = i+9
-        dinput['x'][counter] = np.float64(line[i:j])
-        i = j+10; j = i+9
-        dinput['y'][counter] = np.float64(line[i:j])
-        counter += 1
-    # reduce to data values
-    dinput['MJD'] = dinput['MJD'][:counter]
-    dinput['x'] = dinput['x'][:counter]
-    dinput['y'] = dinput['y'][:counter]
-    # return the date, flag and polar motion values
-    return dinput
+    # raise warnings for deprecation of module
+    warnings.filterwarnings("always")
+    warnings.warn("Deprecated. Please use pyTMD.eop instead",DeprecationWarning)
+    # call renamed version to not break workflows
+    return pyTMD.eop.iers_daily_EOP(input_file)
