+++ pyTMD/read_ocean_pole_tide.py
@@ -1,6 +1,8 @@
 #!/usr/bin/env python
 u"""
-read_ocean_pole_tide.py (04/2022)
+read_ocean_pole_tide.py
+Written by Tyler Sutterley (12/2022)
+
 Reads ocean pole load tide coefficients provided by IERS
 http://maia.usno.navy.mil/conventions/2010/2010_official/chapter7/tn36_c7.pdf
 http://maia.usno.navy.mil/conventions/2010/2010_update/chapter7/icc7.pdf
@@ -29,6 +31,7 @@ REFERENCES:
         doi: 10.1007/s00190-015-0848-7
 
 UPDATE HISTORY:
+    Updated 12/2022: refactored ocean pole tide read program under io
     Updated 04/2022: updated docstrings to numpy documentation format
         use longcomplex data format to be windows compliant
     Updated 07/2021: added check that ocean pole tide file is accessible
@@ -38,10 +41,8 @@ UPDATE HISTORY:
     Updated 12/2018: Compatibility updates for Python3
     Written 09/2017
 """
-import os
-import re
-import gzip
-import numpy as np
+import warnings
+import pyTMD.io
 
 # PURPOSE: read real and imaginary ocean pole tide coefficients
 def read_ocean_pole_tide(input_file):
@@ -74,91 +75,7 @@ def read_ocean_pole_tide(input_file):
         satellite altimetry", Journal of Geodesy, 89(12), p1233-1243, 2015.
         doi: 10.1007/s00190-015-0848-7
     """
-    # check that ocean pole tide file is accessible
-    if not os.access(os.path.expanduser(input_file), os.F_OK):
-        raise FileNotFoundError(os.path.expanduser(input_file))
-
-    # read GZIP ocean pole tide file
-    with gzip.open(os.path.expanduser(input_file),'rb') as f:
-        file_contents = f.read().splitlines()
-
-    # counts the number of lines in the header
-    count = 0
-    # Reading over header text
-    HEADER = True
-    while HEADER:
-        # file line at count
-        line = file_contents[count]
-        # find --------- within line to set HEADER flag to False when found
-        HEADER = not bool(re.match(b'---------',line))
-        # add 1 to counter
-        count += 1
-
-    # grid parameters and dimensions
-    dlon,dlat = (0.50,0.50)
-    glon = np.arange(dlon/2.0,360+dlon/2.0,dlon)
-    glat = np.arange(90.0-dlat/2.0,-90.0-dlat/2.0,-dlat)
-    nlon = len(glon)
-    nlat = len(glat)
-    # allocate for output grid maps
-    ur = np.zeros((nlon,nlat),dtype=np.longcomplex)
-    un = np.zeros((nlon,nlat),dtype=np.longcomplex)
-    ue = np.zeros((nlon,nlat),dtype=np.longcomplex)
-    # read lines of file and add to output variables
-    for i,line in enumerate(file_contents[count:]):
-        ln,lt,urr,uri,unr,uni,uer,uei = np.array(line.split(), dtype='f8')
-        ilon = int(ln/dlon)
-        ilat = int((90.0-lt)/dlat)
-        ur[ilon,ilat] = urr + 1j*uri
-        un[ilon,ilat] = unr + 1j*uni
-        ue[ilon,ilat] = uer + 1j*uei
-
-    # extend matrix for bilinear interpolation
-    glon = extend_array(glon,dlon)
-    ur = extend_matrix(ur)
-    un = extend_matrix(un)
-    ue = extend_matrix(ue)
-    # return values
-    return (ur,un,ue,glon,glat)
-
-# PURPOSE: wrapper function to extend an array
-def extend_array(input_array,step_size):
-    """
-    Wrapper function to extend an array
-
-    Parameters
-    ----------
-    input_array: array to extend
-    step_size: step size between elements of array
-
-    Returns
-    -------
-    temp: extended array
-    """
-    n = len(input_array)
-    temp = np.zeros((n+2),dtype=input_array.dtype)
-    # extended array [x-1,x0,...,xN,xN+1]
-    temp[0] = input_array[0] - step_size
-    temp[1:-1] = input_array[:]
-    temp[-1] = input_array[-1] + step_size
-    return temp
-
-# PURPOSE: wrapper function to extend a matrix
-def extend_matrix(input_matrix):
-    """
-    Wrapper function to extend a matrix
-
-    Parameters
-    ----------
-    input_matrix: matrix to extend
-
-    Returns
-    -------
-    temp: extended matrix
-    """
-    nx,ny = np.shape(input_matrix)
-    temp = np.zeros((nx+2,ny),dtype=input_matrix.dtype)
-    temp[0,:] = input_matrix[-1,:]
-    temp[1:-1,:] = input_matrix[:,:]
-    temp[-1,:] = input_matrix[0,:]
-    return temp
+    # raise warnings for deprecation of module
+    warnings.filterwarnings("always")
+    warnings.warn("Deprecated. Please use pyTMD.io instead",DeprecationWarning)
+    pyTMD.io.ocean_pole_tide(input_file)
