+++ test/test_download_and_read.py
@@ -52,7 +52,6 @@ import pyTMD.model
 import pyTMD.utilities
 import pyTMD.read_tide_model
 import pyTMD.predict_tidal_ts
-import pyTMD.infer_minor_corrections
 import pyTMD.check_tide_points
 import pyTMD.tidal_ellipse
 from oct2py import octave
@@ -209,13 +208,13 @@ class Test_CATS2008:
         assert (hz.shape == (ny,nx))
         assert (mz.shape == (ny,nx))
         # check constituent list
-        constituents,nc = pyTMD.read_tide_model.read_constituents(elevation_file)
+        constituents,nc = pyTMD.io.OTIS.read_constituents(elevation_file)
         cons = ['m2','s2','n2','k2','k1','o1','p1','q1','mf','mm']
         assert all(c in constituents for c in cons)
         # check dimensions of input grids from elevation and transport files
         for i,c in enumerate(constituents):
-            z = pyTMD.read_tide_model.read_elevation_file(elevation_file,i)
-            u,v = pyTMD.read_tide_model.read_transport_file(transport_file,i)
+            z = pyTMD.io.OTIS.read_otis_elevation(elevation_file,i)
+            u,v = pyTMD.io.OTIS.read_otis_transport(transport_file,i)
             assert (z.shape == (ny,nx))
             assert (u.shape == (ny,nx))
             assert (v.shape == (ny,nx))
@@ -283,7 +282,7 @@ class Test_CATS2008:
         station_ph.data[station_ph.mask] = station_ph.fill_value
 
         # extract amplitude and phase from tide model
-        amp,ph,D,cons = pyTMD.read_tide_model.extract_tidal_constants(station_lon,
+        amp,ph,D,cons = pyTMD.io.OTIS.extract_constants(station_lon,
             station_lat, grid_file, model_file, EPSG, type=TYPE, method='spline',
             grid=GRID)
         # reorder constituents of model and convert amplitudes to cm
@@ -424,9 +423,9 @@ class Test_CATS2008:
             tide.mask = np.zeros((ndays),dtype=bool)
             # predict tidal elevations at time and infer minor corrections
             tide.mask[:] = np.any(hc.mask)
-            tide.data[:] = pyTMD.predict_tidal_ts(tide_time, hc, c,
+            tide.data[:] = pyTMD.predict.time_series(tide_time, hc, c,
                 deltat=deltat, corrections=GRID)
-            minor = pyTMD.infer_minor_corrections(tide_time, hc, c,
+            minor = pyTMD.predict.infer_minor(tide_time, hc, c,
                 deltat=deltat, corrections=GRID)
             tide.data[:] += minor.data[:]
 
@@ -731,9 +730,9 @@ class Test_AOTIM5_2018:
             tide.mask = np.zeros((ndays),dtype=bool)
             # predict tidal elevations at time and infer minor corrections
             tide.mask[:] = np.any(hc.mask)
-            tide.data[:] = pyTMD.predict_tidal_ts(tide_time, hc, c,
+            tide.data[:] = pyTMD.predict.time_series(tide_time, hc, c,
                 deltat=deltat, corrections=GRID)
-            minor = pyTMD.infer_minor_corrections(tide_time, hc, c,
+            minor = pyTMD.predict.infer_minor(tide_time, hc, c,
                 deltat=deltat, corrections=GRID)
             tide.data[:] += minor.data[:]
